refactor: move common aoc functions to, #44 * /aoc/file - file handlers * /aoc/array - array handlers * /aoc/grid - common aoc grid input * /aoc/number - number utilities * /aoc/point - common aoc grid-coordinate handlers
