+++ src/2024/2024-12-03/main.ts
@@ -1,5 +1,5 @@
 import path from 'path'
-import { directory, readFile } from '@/aoc/file.js'
+import { directory, readFile } from '@/aoc/file/utils.js'
 import { extractMultiply, extractMultiplyCondition } from './lib/extractMultiply.js'
 
 const dir = directory(import.meta.url)
