+++ src/2024/2024-12-05/lib/orderedUpdates.ts
@@ -1,5 +1,5 @@
 import type { QuizData, Rules } from './fileReader.js'
-import { uniformArrayElements, arrayMiddleIndex } from '@/aoc/arrays.js'
+import { arrayMiddleIndex, uniformArrayElements } from '@/aoc/array/utils.js'
 
 /**
  * Checks if an "update" list is correct according to defined "rules"
