--- src/2024/2024-12-06/lib/grid.types.ts
@@ -1,10 +1,0 @@
-/**
- * `Grid` dimensions properties
- *
- * @property {number} length - Length of the grid
- * @property {number} wwidth - Width of the grid
- */
-export type GridDimensions = {
-  length: number;
-  width: number;
-}
