+++ src/2024/2024-12-09/main.ts
@@ -1,7 +1,7 @@
 import path from 'path'
 
-import { AOC_OUTPUT_TYPE, readAOCInputFile } from '@/aoc/aocInputFile.js'
-import { directory } from '@/aoc/file.js'
+import { AOC_OUTPUT_TYPE, readAOCInputFile } from '@/aoc/file/aocfile.js'
+import { directory } from '@/aoc/file/utils.js'
 
 import { CompactDisk } from './lib/compact.js'
 import { WholeDisk } from './lib/whole.js'
