+++ src/2024/2024-12-10/README.md
@@ -13,6 +13,6 @@ Visit the Advent of Code website for more information on this puzzle at:
 
 ### `scoresRatings.ts`
 
-- `countTrailScores()`
+- **`countTrailScores()`**
    - Finds valid hiking trails (trailheads) from a point coordinate in a 2D array starting with `0` and ending in `9` symbol and calculates the **scores** for each trailhead.
    - Calculates the trailhead **ratings** instead of the trailhead scores if provided with the optional `{ isRating: true }` parameter. Defaults to `false`
