+++ src/2024/2024-12-10/main.ts
@@ -1,6 +1,6 @@
 
-import { readAOCInputFile, AOC_OUTPUT_TYPE } from '@/aoc/aocInputFile.js'
-import { file } from '@/aoc/file.js'
+import { readAOCInputFile, AOC_OUTPUT_TYPE } from '@/aoc/file/aocfile.js'
+import { file } from '@/aoc/file/utils.js'
 
 import { countTrailScores } from './lib/scoresRatings.js'
