+++ src/2024/2024-12-11/README.md
@@ -10,9 +10,7 @@ Visit the Advent of Code website for more information on this puzzle at:
 ### `utils.ts`
 
 - contains helper and utility scripts for the quiz.
-- `isEvenDigits()` - Checks if a number has an even number of digits
-- `halfDigit()` - Divides the number into two (2) separate groups (digits) if they have an even number of digits, each group having half the original number's digits.
-- `arrayToObject()` - Converts an array of numbers to an Object, using the array number elements as keys with a default numeric value of `1`
+- **`halfDigit()`** - Divides the number into two (2) separate groups (digits) if they have an even number of digits, each group having half the original number's digits.
 
 ### `blink.ts`
