+++ src/2024/2024-12-12/README.md
@@ -9,27 +9,22 @@ Visit the Advent of Code website for more information on this puzzle at:
 
 ### `garden.ts`
 
-- **Garden** class
-   - A set of methods and properties for calculating Garden, Region, and Plots data
-   - `Garden.calculatePlot()` - Calculates the per-plot perimeter and total area of all garden regions in a 2D grid, each defined by an initial symbol at a starting (y,x) coordinate.
-   - `Garden.calculateFencePrice()` - Calculates the total fencing price of all regions in a garden per connected plot using the formula: area * perimeter (per plot).
+**Garden** class
+- A set of methods and properties for calculating Garden, Region, and Plots data
+- `Garden.calculatePlot()` - Calculates the per-plot perimeter and total area of all garden regions in a 2D grid, each defined by an initial symbol at a starting (y,x) coordinate.
+- `Garden.calculateFencePrice()` - Calculates the total fencing price of all regions in a garden per connected plot using the formula: area * perimeter (per plot).
 
 ### `wholesale.ts`
 
-- **WholesaleGarden** class
-   - A set of methods and properties for calculating the wholesale fencing price of garden regions
-   - `WholesaleGarden.calculateRegionCorners()` - Calculates the number of corners (sides) of a whole region.
-   - `WholesaleGarden.calculateFencePrice()` - Calculates the total fencing price of all regions in a garden using the formula: area * perimeter (of whole region).
+**WholesaleGarden** class
+- A set of methods and properties for calculating the wholesale fencing price of garden regions
+- **`calculateRegionCorners()`** - Calculates the number of corners (sides) of a whole region.
+- **`calculateFencePrice()`** - Calculates the total fencing price of all regions in a garden using the formula: area * perimeter (of whole region).
 
 ### `utils.ts`
 
 - A script containing helper and utility scripts for the quiz
-- `findNeighbors()` - Finds all coordinates of the neighboring plots from a specified coordinate (up/down/left/right)
-- `isIllegalCoordinate()` - Checks if a point at coordinate (y,x) in a grid is illegal: if it's out of bounds of the grid area or if its symbol differs from the symbol parameter.
-- `getDiagonalNeighbors()` - Retrieves the four (4) diagonally-aligned (y,x) coordinates and the symbol character from a `Point` in the grid. Substitutes a `"*"` symbol character in the `NeighborPoint.symbol`if the `point` is out of the grid bounds.
-- `getCrossNeighbors()` - Retrieves the four (4) horizontal/vertical aligned (y,x) coordinates and the symbol character from a `Point` in the grid. Substitutes a `"*"` symbol character in the `NeighborPoint.symbol`if the `point` is out of the grid bounds.
-- `isDiagonal()` - Checks if two (2) `Points` are diagonally-aligned
-- `innerCorners()` - Counts the "inner" corners from groups of valid L-shaped `Points` that has the `point` parameter coordinate as its center
+- **`innerCorners()`** - Counts the "inner" corners from groups of valid L-shaped `Points` that has the `point` parameter coordinate as its center
 
 
 ## Notes
