+++ src/2024/2024-12-12/sample.test.ts
@@ -1,8 +1,8 @@
 import { test, expect } from 'vitest'
-import { AOC_OUTPUT_TYPE } from '@/aoc/aocInputFile.js'
-import { file } from '@/aoc/file.js'
+import { AOC_OUTPUT_TYPE } from '@/aoc/file/aocfile.js'
+import { file } from '@/aoc/file/utils.js'
 
-import { readAOCInputFileAsync } from '@/aoc/aocInputFile.js'
+import { readAOCInputFileAsync } from '@/aoc/file/aocfile.js'
 
 import { Garden } from './lib/garden.js'
 import { WholesaleGarden } from './lib/wholesale.js'
