+++ src/aoc/array/utils.ts
@@ -20,7 +20,30 @@ export const arrangeArray = (order: ARRAY_ORDERING) =>
   }
 
 /**
- * Checks if array elements have the same type using `typeof` and has no null or undefined values
+ * Converts an array of numbers to an Object, using the array number elements as keys with a default numeric value of `1`
+ * @param {number[]} numbers - Array of numbers
+ * @returns {Record<number, number>} Object
+ */
+export const arrayToObject = (numbers: number[]): Record<number, number> => {
+  return numbers.reduce((list: Record<number, number>, num) =>
+    (list[num] === undefined
+      ? list[num] = 1
+      : list[num] += 1, list),
+  {}
+  )
+}
+
+/**
+ * Retrieves the middle (center) index of an array
+ * @param list {T[]} 1-dimensional array
+ * @returns {number} Middle index of an array
+ */
+export const arrayMiddleIndex = <T>(list: T[]): number => {
+  return Math.floor(list.length / 2) + list.length % 2
+}
+
+/**
+ * Checks if array elements have the same type using `typeof` and have no null or undefined values
  * @param items {S[]} array of elements
  * @param type {T} primitive type name of the elements inside the array (e.g., number, string, boolean)
  * @returns {boolean} Flag indicating if all array elements have the same type
@@ -33,13 +56,4 @@ export const uniformArrayElements = <S, T>(
     items.filter(value => typeof value === type)
       .length === items.length
   )
-}
-
-/**
- * Retrieves the middle (center) index of an array
- * @param list {T[]} 1-dimensional array
- * @returns {number} Middle index of an array
- */
-export const arrayMiddleIndex = <T>(list: T[]): number => {
-  return Math.floor(list.length / 2) + list.length % 2
 }
