+++ src/aoc/number/utils.ts
@@ -11,3 +11,11 @@ export const isNumber = (input: string | number): boolean => {
 
   return !isNaN(Number(input))
 }
+
+/**
+ * Checks if a number has an even number of digits
+ * @param {number} number - Number input
+ */
+export const isEvenDigits = (number: number): boolean => {
+  return number.toString().length % 2 === 0
+}
